refactor feedbacks 24x7 and midsem
